feat: refactor overlay to be reusable
